+++ command/build.go
@@ -6,17 +6,13 @@ import (
 	"fmt"
 	"log"
 	"math"
-	"os"
-	"os/signal"
 	"strconv"
 	"strings"
 	"sync"
-	"syscall"
 
 	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/hclparse"
 	"github.com/hashicorp/packer/hcl2template"
-	"github.com/hashicorp/packer/helper/enumflag"
 	"github.com/hashicorp/packer/packer"
 	"github.com/hashicorp/packer/template"
 	"golang.org/x/sync/semaphore"
@@ -29,60 +25,26 @@ type BuildCommand struct {
 }
 
 func (c *BuildCommand) Run(args []string) int {
-	buildCtx, cancelBuildCtx := context.WithCancel(context.Background())
-	// Handle interrupts for this build
-	sigCh := make(chan os.Signal, 1)
-	signal.Notify(sigCh, os.Interrupt, syscall.SIGTERM)
-	defer func() {
-		cancelBuildCtx()
-		signal.Stop(sigCh)
-		close(sigCh)
-	}()
-	go func() {
-		select {
-		case sig := <-sigCh:
-			if sig == nil {
-				// context got cancelled and this closed chan probably
-				// triggered first
-				return
-			}
-			c.Ui.Error(fmt.Sprintf("Cancelling build after receiving %s", sig))
-			cancelBuildCtx()
-		case <-buildCtx.Done():
-		}
-	}()
+	ctx, cleanup := handleTermInterrupt(c.Ui)
+	defer cleanup()
 
-	return c.RunContext(buildCtx, args)
+	cfg, ret := c.ParseArgs(args)
+	if ret != 0 {
+		return ret
+	}
+
+	return c.RunContext(ctx, cfg)
 }
 
-// Config is the command-configuration parsed from the command line.
-type Config struct {
-	Color, Debug, Force, Timestamp bool
-	ParallelBuilds                 int64
-	OnError                        string
-	Path                           string
-}
-
-func (c *BuildCommand) ParseArgs(args []string) (Config, int) {
-	var cfg Config
-	var parallel bool
+func (c *BuildCommand) ParseArgs(args []string) (*BuildArgs, int) {
+	var cfg BuildArgs
 	flags := c.Meta.FlagSet("build", FlagSetBuildFilter|FlagSetVars)
 	flags.Usage = func() { c.Ui.Say(c.Help()) }
-	flags.BoolVar(&cfg.Color, "color", true, "")
-	flags.BoolVar(&cfg.Debug, "debug", false, "")
-	flags.BoolVar(&cfg.Force, "force", false, "")
-	flags.BoolVar(&cfg.Timestamp, "timestamp-ui", false, "")
-	flagOnError := enumflag.New(&cfg.OnError, "cleanup", "abort", "ask")
-	flags.Var(flagOnError, "on-error", "")
-	flags.BoolVar(&parallel, "parallel", true, "")
-	flags.Int64Var(&cfg.ParallelBuilds, "parallel-builds", 0, "")
+	cfg.AddFlagSets(flags)
 	if err := flags.Parse(args); err != nil {
-		return cfg, 1
+		return &cfg, 1
 	}
 
-	if parallel == false && cfg.ParallelBuilds == 0 {
-		cfg.ParallelBuilds = 1
-	}
 	if cfg.ParallelBuilds < 1 {
 		cfg.ParallelBuilds = math.MaxInt64
 	}
@@ -90,101 +52,98 @@ func (c *BuildCommand) ParseArgs(args []string) (Config, int) {
 	args = flags.Args()
 	if len(args) != 1 {
 		flags.Usage()
-		return cfg, 1
+		return &cfg, 1
 	}
 	cfg.Path = args[0]
-	return cfg, 0
+	return &cfg, 0
 }
 
-func (c *BuildCommand) GetBuildsFromHCL(path string) ([]packer.Build, int) {
+func (m *Meta) GetConfigFromHCL(cla *MetaArgs) (packer.BuildGetter, int) {
 	parser := &hcl2template.Parser{
 		Parser:                hclparse.NewParser(),
-		BuilderSchemas:        c.CoreConfig.Components.BuilderStore,
-		ProvisionersSchemas:   c.CoreConfig.Components.ProvisionerStore,
-		PostProcessorsSchemas: c.CoreConfig.Components.PostProcessorStore,
+		BuilderSchemas:        m.CoreConfig.Components.BuilderStore,
+		ProvisionersSchemas:   m.CoreConfig.Components.ProvisionerStore,
+		PostProcessorsSchemas: m.CoreConfig.Components.PostProcessorStore,
 	}
 
-	builds, diags := parser.Parse(path, c.varFiles, c.flagVars, c.CoreConfig.Only, c.CoreConfig.Except)
-	{
-		// write HCL errors/diagnostics if any.
-		b := bytes.NewBuffer(nil)
-		err := hcl.NewDiagnosticTextWriter(b, parser.Files(), 80, false).WriteDiagnostics(diags)
-		if err != nil {
-			c.Ui.Error("could not write diagnostic: " + err.Error())
-			return nil, 1
-		}
-		if b.Len() != 0 {
-			c.Ui.Message(b.String())
-		}
-	}
-	ret := 0
-	if diags.HasErrors() {
-		ret = 1
-	}
-
-	return builds, ret
+	cfg, diags := parser.Parse(cla.Path, cla.VarFiles, cla.Vars)
+	return cfg, writeDiags(m.Ui, parser.Files(), diags)
 }
 
-func (c *BuildCommand) GetBuilds(path string) ([]packer.Build, int) {
+func writeDiags(ui packer.Ui, files map[string]*hcl.File, diags hcl.Diagnostics) int {
+	// write HCL errors/diagnostics if any.
+	b := bytes.NewBuffer(nil)
+	err := hcl.NewDiagnosticTextWriter(b, files, 80, false).WriteDiagnostics(diags)
+	if err != nil {
+		ui.Error("could not write diagnostic: " + err.Error())
+		return 1
+	}
+	if b.Len() != 0 {
+		if diags.HasErrors() {
+			ui.Error(b.String())
+			return 1
+		}
+		ui.Say(b.String())
+	}
+	return 0
+}
 
-	isHCLLoaded, err := isHCLLoaded(path)
-	if path != "-" && err != nil {
-		c.Ui.Error(fmt.Sprintf("could not tell whether %s is hcl enabled: %s", path, err))
+func (m *Meta) GetConfig(cla *MetaArgs) (packer.BuildGetter, int) {
+	cfgType, err := ConfigType(cla.Path)
+	if err != nil {
+		m.Ui.Error(fmt.Sprintf("could not tell config type: %s", err))
 		return nil, 1
 	}
-	if isHCLLoaded {
-		return c.GetBuildsFromHCL(path)
+
+	switch cfgType {
+	case "hcl":
+		// TODO(azr): allow to pass a slice of files here.
+		return m.GetConfigFromHCL(cla)
+	default:
+		// TODO: uncomment once we've polished HCL a bit more.
+		// c.Ui.Say(`Legacy JSON Configuration Will Be Used.
+		// The template will be parsed in the legacy configuration style. This style
+		// will continue to work but users are encouraged to move to the new style.
+		// See: https://packer.io/guides/hcl
+		// `)
+		return m.GetConfigFromJSON(cla)
 	}
+}
 
-	// TODO: uncomment in v1.5.1 once we've polished HCL a bit more.
-	// c.Ui.Say(`Legacy JSON Configuration Will Be Used.
-	// The template will be parsed in the legacy configuration style. This style
-	// will continue to work but users are encouraged to move to the new style.
-	// See: https://packer.io/guides/hcl
-	// `)
-
+func (m *Meta) GetConfigFromJSON(cla *MetaArgs) (packer.BuildGetter, int) {
 	// Parse the template
-	var tpl *template.Template
-	tpl, err = template.ParseFile(path)
+	tpl, err := template.ParseFile(cla.Path)
 	if err != nil {
-		c.Ui.Error(fmt.Sprintf("Failed to parse template: %s", err))
+		m.Ui.Error(fmt.Sprintf("Failed to parse template: %s", err))
 		return nil, 1
 	}
 
 	// Get the core
-	core, err := c.Meta.Core(tpl)
-	if err != nil {
-		c.Ui.Error(err.Error())
-		return nil, 1
-	}
-
+	core, err := m.Core(tpl, cla)
 	ret := 0
-	buildNames := c.Meta.BuildNames(core)
-	builds := make([]packer.Build, 0, len(buildNames))
-	for _, n := range buildNames {
-		b, err := core.Build(n)
-		if err != nil {
-			c.Ui.Error(fmt.Sprintf(
-				"Failed to initialize build '%s': %s",
-				n, err))
-			ret = 1
-			continue
-		}
-
-		builds = append(builds, b)
+	if err != nil {
+		m.Ui.Error(err.Error())
+		ret = 1
 	}
-	return builds, ret
+	return core, ret
 }
 
-func (c *BuildCommand) RunContext(buildCtx context.Context, args []string) int {
-	cfg, ret := c.ParseArgs(args)
+func (c *BuildCommand) RunContext(buildCtx context.Context, cla *BuildArgs) int {
+	packerStarter, ret := c.GetConfig(&cla.MetaArgs)
 	if ret != 0 {
 		return ret
 	}
 
-	builds, ret := c.GetBuilds(cfg.Path)
+	builds, diags := packerStarter.GetBuilds(packer.GetBuildsOptions{
+		Only:   cla.Only,
+		Except: cla.Except,
+	})
 
-	if cfg.Debug {
+	// here, something could have gone wrong but we still want to run valid
+	// builds.
+	ret = writeDiags(c.Ui, nil, diags)
+
+	if cla.Debug {
 		c.Ui.Say("Debug mode enabled. Builds will not be parallelized.")
 	}
 
@@ -199,7 +158,7 @@ func (c *BuildCommand) RunContext(buildCtx context.Context, args []string) int {
 	buildUis := make(map[packer.Build]packer.Ui)
 	for i := range builds {
 		ui := c.Ui
-		if cfg.Color {
+		if cla.Color {
 			ui = &packer.ColoredUi{
 				Color: colors[i%len(colors)],
 				Ui:    ui,
@@ -213,7 +172,7 @@ func (c *BuildCommand) RunContext(buildCtx context.Context, args []string) int {
 			}
 		}
 		// Now add timestamps if requested
-		if cfg.Timestamp {
+		if cla.TimestampUi {
 			ui = &packer.TimestampedUi{
 				Ui: ui,
 			}
@@ -222,17 +181,17 @@ func (c *BuildCommand) RunContext(buildCtx context.Context, args []string) int {
 		buildUis[builds[i]] = ui
 	}
 
-	log.Printf("Build debug mode: %v", cfg.Debug)
-	log.Printf("Force build: %v", cfg.Force)
-	log.Printf("On error: %v", cfg.OnError)
+	log.Printf("Build debug mode: %v", cla.Debug)
+	log.Printf("Force build: %v", cla.Force)
+	log.Printf("On error: %v", cla.OnError)
 
 	// Set the debug and force mode and prepare all the builds
 	for i := range builds {
 		b := builds[i]
 		log.Printf("Preparing build: %s", b.Name())
-		b.SetDebug(cfg.Debug)
-		b.SetForce(cfg.Force)
-		b.SetOnError(cfg.OnError)
+		b.SetDebug(cla.Debug)
+		b.SetForce(cla.Force)
+		b.SetOnError(cla.OnError)
 
 		warnings, err := b.Prepare()
 		if err != nil {
@@ -260,7 +219,7 @@ func (c *BuildCommand) RunContext(buildCtx context.Context, args []string) int {
 		sync.RWMutex
 		m map[string]error
 	}{m: make(map[string]error)}
-	limitParallel := semaphore.NewWeighted(cfg.ParallelBuilds)
+	limitParallel := semaphore.NewWeighted(cla.ParallelBuilds)
 	for i := range builds {
 		if err := buildCtx.Err(); err != nil {
 			log.Println("Interrupted, not going to start any more builds.")
@@ -304,12 +263,12 @@ func (c *BuildCommand) RunContext(buildCtx context.Context, args []string) int {
 			}
 		}()
 
-		if cfg.Debug {
+		if cla.Debug {
 			log.Printf("Debug enabled, so waiting for build to finish: %s", b.Name())
 			wg.Wait()
 		}
 
-		if cfg.ParallelBuilds == 1 {
+		if cla.ParallelBuilds == 1 {
 			log.Printf("Parallelization disabled, waiting for build to finish: %s", b.Name())
 			wg.Wait()
 		}
@@ -415,8 +374,7 @@ Options:
   -force                        Force a build to continue if artifacts exist, deletes existing artifacts.
   -machine-readable             Produce machine-readable output.
   -on-error=[cleanup|abort|ask] If the build fails do: clean up (default), abort, or ask.
-  -parallel=false               Disable parallelization. (Default: true)
-  -parallel-builds=1            Number of builds to run in parallel. 0 means no limit (Default: 0)
+  -parallel-builds=1            Number of builds to run in parallel. 1 disables parallelization. 0 means no limit (Default: 0)
   -timestamp-ui                 Enable prefixing of each ui output with an RFC3339 timestamp.
   -var 'key=value'              Variable for templates, can be used multiple times.
   -var-file=path                JSON file containing user variables.