+++ hcl2template/parser.go
@@ -52,7 +52,15 @@ const (
 	hcl2VarJsonFileExt = ".auto.pkrvars.json"
 )
 
-func (p *Parser) parse(filename string, varFiles []string, argVars map[string]string) (*PackerConfig, hcl.Diagnostics) {
+// Parse will Parse all HCL files in filename. Path can be a folder or a file.
+//
+// Parse will first Parse variables and then the rest; so that interpolation
+// can happen.
+//
+// Parse returns a PackerConfig that contains configuration layout of a packer
+// build; sources(builders)/provisioners/posts-processors will not be started
+// and their contents wont be verified; Most syntax errors will cause an error.
+func (p *Parser) Parse(filename string, varFiles []string, argVars map[string]string) (*PackerConfig, hcl.Diagnostics) {
 
 	var files []*hcl.File
 	var diags hcl.Diagnostics
@@ -89,7 +97,10 @@ func (p *Parser) parse(filename string, varFiles []string, argVars map[string]st
 		basedir = filepath.Dir(basedir)
 	}
 	cfg := &PackerConfig{
-		Basedir: basedir,
+		Basedir:               basedir,
+		builderSchemas:        p.BuilderSchemas,
+		provisionersSchemas:   p.ProvisionersSchemas,
+		postProcessorsSchemas: p.PostProcessorsSchemas,
 	}
 
 	// Decode variable blocks so that they are available later on. Here locals